% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_scrapeGS.R
\name{crossref_lookup}
\alias{crossref_lookup}
\title{Global wrapper for CrossRef lookup based on scraped Google Scholar results}
\usage{
crossref_lookup(data)
}
\arguments{
\item{data}{dataframe output object from `save_and_scrapeGS()`, containing digital
object identifiers (DOIs; column DO) and article titles (column TI).}
}
\value{
A list of (lists of) CrossRef citations for all retrieved records.
}
\description{
Function wraps doi and title lookups for scraped results from Google
Scholar results pages.
}
\examples{
crossref_results <- crossref_lookup(info)
head(crossref_results);
}
