% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_dois.R
\name{decode_dois}
\alias{decode_dois}
\title{Decode DOIs (digital object identifiers)}
\usage{
decode_dois(dois, keep_http = TRUE)
}
\arguments{
\item{dois}{A vector of DOIs.}

\item{keep_http}{Logical argument specifying whether to retain http
prefix (TRUE) or not (FALSE). Default is set to TRUE.}
}
\value{
A vector of edited DOIs
}
\description{
Replace URL encoded DOIs (often exported from bibliographic
databases and search tools like Publish or Perish, e.g. '%28' = '(')
with original special character encoding (e.g. '(').
}
\examples{
\dontrun{
refs <- synthesisr::read_refs('www/embase.ris')
dois <- refs$doi
new_dois <- decode_dois(dois, keep_http = FALSE)
}
}
