% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{get_authors}
\alias{get_authors}
\title{Extract authors from Google Scholar results}
\usage{
get_authors(html)
}
\arguments{
\item{html}{A vector of lines consisting of the html code for a web page}
}
\value{
A vector of author lists
}
\description{
Extract author details from the record's citation from Google
Scholar search results.
}
\examples{
authors <- get_authors(lines)
authors;
}
