% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{get_html_code}
\alias{get_html_code}
\title{Scrape in code from local html file}
\usage{
get_html_code(filename)
}
\arguments{
\item{filename}{Name of the file to be scraped. File should be saved in the working
directory.}
}
\value{
A string of html code for a webpage
}
\description{
Scrape in code from an html file saved locally
}
\examples{
code <- get_html_code(file.choose());
}
