% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{get_htmls_code}
\alias{get_htmls_code}
\title{Scrape in code from multiple local html files}
\usage{
get_htmls_code()
}
\value{
One or more strings of html code for a webpage, stored as a list
}
\description{
Scrape in code from html files saved locally. Files should be saved in the working
directory.
}
\examples{
html_codes <- get_htmls_code()
file1 <- unlist(html_codes[1]);
}
