% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{get_links}
\alias{get_links}
\title{Extract article links from Google Scholar results}
\usage{
get_links(code_lines)
}
\arguments{
\item{code_lines}{A vector of lines consisting of the html code for a webpage}
}
\value{
A vector of URLs
}
\description{
Extract links to websites holding article information from Google
Scholar search results.
}
\examples{
\dontrun{
links <- get_links(code_lines)
links
}
}
