% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{get_titles}
\alias{get_titles}
\title{Extract titles from Google Scholar results}
\usage{
get_titles(html)
}
\arguments{
\item{html}{A vector of lines consisting of the html code for a webpage}
}
\value{
A vector of result titles (10 per page)
}
\description{
Extract the titles of search results from Google Scholar search results, saved as
html files.
}
\examples{
titles <- get_titles(lines);
}
