% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{links2dois}
\alias{links2dois}
\title{Extract DOIs from multiple article links}
\usage{
links2dois(code_lines)
}
\arguments{
\item{code_lines}{A data frame of links to a webpage from Google Scholar}
}
\value{
A data frame of dois
}
\description{
Extract digital object identifiers ('DOIs') from  links to the websites
holding article information, where they are contained within (e.g.
'https://link.springer.com/article/10.1186/1475-2875-13-446' contains the DOI
10.1186/1475-2875-13-446).
}
\examples{
\dontrun{
dois <- links2dois(code_lines)
}
}
