% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSscraper.R
\name{runGSscraper}
\alias{runGSscraper}
\title{Wrapper to run GSscraper Shiny app locally}
\usage{
runGSscraper()
}
\value{
Runs the GitHub version of the Shiny App locally.
}
\description{
Run the GSscraper Shiny app locally with this
function. Calls the latest version of the app from GitHub without
the need to install GSscraper.
}
\examples{
\dontrun{
runGSscraper()
}
}
