% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_html}
\alias{save_html}
\title{Download one results page as html file}
\usage{
save_html(url, sleep = "4")
}
\arguments{
\item{url}{One URLs corresponding to a page of Google Scholar search results.}

\item{sleep}{Integer specifying the number of seconds to wait between download attempts. The
default value is 4 seconds.}
}
\value{
An html file is downloaded with a file name corresponding to the URL with punctuation removed
for clarity. Files are saved to the working directory. A pause notification is printed to the
console.
}
\description{
Download one results page as html file
}
\examples{
urls <- 'https://scholar.google.co.uk/scholar?start=50&q=insect+population+\%22systematic+review\%22&hl=en&as_vis=0,5&as_sdt=0,5'
save_html(urls, sleep = '5');
}
