% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{split_by_div}
\alias{split_by_div}
\title{Split file by '<div' code into different lines}
\usage{
split_by_div(html)
}
\arguments{
\item{html}{A string consisting of the html code for a webpage}
}
\value{
A vector of strings, one for each separated line
}
\description{
Split the imported html code into lines based on the '<div' field code.
}
\examples{
\dontrun{
lines <- split_by_div(code)
lines
}
}
