% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{get_citations}
\alias{get_citations}
\title{Extract citation information from Google Scholar results}
\usage{
get_citations(code_lines)
}
\arguments{
\item{code_lines}{A vector of lines consisting of the html code for a webpage}
}
\value{
A vector of citation information
}
\description{
Extract information relating to the record's citation (including the authors,
journal/source and publication year), from Google Scholar search results.
}
\examples{
\dontrun{
citations <- get_citations(code_lines)
}
}
