% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{get_descriptions}
\alias{get_descriptions}
\title{Extract article descriptions from Google Scholar results}
\usage{
get_descriptions(html)
}
\arguments{
\item{html}{A vector of lines consisting of the html code for a webpage}
}
\value{
A vector of descriptions
}
\description{
Extract article descriptions (summaries of key sentences), from Google
Scholar search results.
}
\examples{
descriptions <- get_descriptions(lines)
descriptions;
}
