% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{get_info}
\alias{get_info}
\title{Scrape all Google Scholar information}
\usage{
get_info(html)
}
\arguments{
\item{html}{An HTML file read in as a single character string using 'save_htmls()'.}
}
\value{
A data frame containing all information that can be extracted from all html files
in the working directory.
}
\description{
A wrapper for the scraping functions to produce a data frame of citations for
each page of Google Scholar search results.
}
\examples{
\dontrun{
info <- get_info(html)
}
}
