% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{get_links}
\alias{get_links}
\title{Extract article links from Google Scholar results}
\usage{
get_links(html)
}
\arguments{
\item{html}{A vector of lines consisting of the html code for a webpage}
}
\value{
A vector of URLs
}
\description{
Extract links to websites holding article information from Google
Scholar search results.
}
\examples{
links <- get_links(lines)
links;
}
