% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{get_years}
\alias{get_years}
\title{Extract article descriptions from Google Scholar results}
\usage{
get_years(code_lines)
}
\arguments{
\item{code_lines}{A vector of lines consisting of the html code for a webpage}
}
\value{
A vector of publication years
}
\description{
Extract article publication years from Google Scholar search results.
}
\examples{
\dontrun{
years <- get_years(lines)
years
}
}
