% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{link2doi}
\alias{link2doi}
\title{Extract a DOI from an article link}
\usage{
link2doi(link)
}
\arguments{
\item{link}{A vector of lines consisting of the html code for a webpage}
}
\value{
A doi
}
\description{
Extract a digital object identifier ('DOI') from the link to the website
holding article information, where they are contained within (e.g.
'https://link.springer.com/article/10.1186/1475-2875-13-446' contains the DOI
10.1186/1475-2875-13-446).
}
\examples{
link <- 'https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0189268'
doi <- link2doi(link)
doi;
}
