% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGS.R
\name{links2dois}
\alias{links2dois}
\title{Extract DOIs from multiple article links}
\usage{
links2dois(link)
}
\arguments{
\item{html}{A vector of lines consisting of the html code for a webpage}
}
\value{
A vector of DOIs
}
\description{
Extract digital object identifiers ('DOIs') from  links to the websites
holding article information, where they are contained within (e.g.
'https://link.springer.com/article/10.1186/1475-2875-13-446' contains the DOI
10.1186/1475-2875-13-446).
}
\examples{
dois <- links2dois(links)
dois;
}
