% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_htmls.R
\name{save_html}
\alias{save_html}
\title{Download one results page as html file}
\usage{
save_html(url, path = "", pause = 4, backoff = FALSE)
}
\arguments{
\item{url}{One URLs corresponding to a page of Google Scholar search results.}

\item{path}{The path in which the file should be saved. The default is to save in the working directory.}

\item{pause}{Integer specifying the number of seconds to wait between download attempts. The
default value is 4 seconds.}

\item{backoff}{A logical argument (TRUE or FALSE) specifying whether responsive backing-off should be used.
If set to TRUE, the time between calls is varied depending on how long the server takes to respond to the
original request. The responsive back-off time is set to multiple the response time by the `pause` time: i.e.
if the system takes 1.02 seconds to respond and `pause` time is set to 4 seconds, a 4.10 second delay will
be employed before the next call. The default for back-off is `FALSE`.}
}
\value{
An HTML file is downloaded with a file name corresponding to the URL with punctuation removed
for clarity. Files are saved to the working directory. A pause notification is printed to the
console.
}
\description{
Downloads one page of Google Scholar results from a URLs as html files with
a specific wait-time to avoid IP address blocking.
}
\examples{
url <- 'https://scholar.google.co.uk/scholar?start=0&q=insect+population+\%22systematic+review\%22&hl=en&as_vis=0,5&as_sdt=0,5'
save_html(url,
    pause = 5,
    backoff = FALSE);
}
