% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_htmls}
\alias{save_htmls}
\title{Download multiple results pages as html files}
\usage{
save_htmls(url, sleep = "4")
}
\arguments{
\item{url}{One or more URLs corresponding to pages of Google Scholar search results.}

\item{sleep}{Integer specifying the number of seconds to wait between download attempts. The
default value is 4 seconds.}
}
\value{
Files are downloaded with a file name corresponding to the URL with punctuation removed
for clarity. Files are saved to the working directory. A pause notification is printed to the
console.
}
\description{
Download multiple results pages as html files
}
\examples{
urls <- c('https://scholar.google.co.uk/scholar?start=50&q=insect+population+\%22systematic+review\%22&hl=en&as_vis=0,5&as_sdt=0,5',
   'https://scholar.google.co.uk/scholar?start=90&q=insect+population+\%22systematic+review\%22&hl=en&as_vis=0,5&as_sdt=0,5')
save_html(urls, sleep = '5');
}
